staking finalized

+++ contracts/Chromium.sol
@@ -2,28 +2,34 @@
 pragma solidity >=0.4.22 <0.9.0;
 
 import "./interfaces/UniversalERC20.sol";
-import './interfaces/Ownable.sol';
-import './ExchangeOracle.sol';
-import './interfaces/IUniswap.sol';
+import "./interfaces/Ownable.sol";
+import "./ExchangeOracle.sol";
+import "./interfaces/IUniswap.sol";
 
 contract Chromium is Ownable {
     using UniversalERC20 for IERC20;
 
     // used to keep track of tokens in contract
-    mapping(IERC20 => uint) public liquidityAmount;
+    mapping(IERC20 => uint256) public liquidityAmount;
 
     // weth address that i am using for eth
-    IERC20 private constant ETH_ADDRESS = IERC20(0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE);
+    IERC20 private constant ETH_ADDRESS =
+        IERC20(0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE);
     // IERC20 private constant cbltToken = IERC20(0x433C6E3D2def6E1fb414cf9448724EFB0399b698);
 
     // initializing objects
     ExchangeOracle oracle;
     IERC20 cbltToken;
 
-    mapping(IERC20 => uint) feeTotal;
+    mapping(IERC20 => uint256) feeTotal;
 
     // emits when chromium is used
-    event ChromiumTrade(address indexed _from, address _fromToken, uint256 _fromAmount, uint _cbltAmount);
+    event ChromiumTrade(
+        address indexed _from,
+        address _fromToken,
+        uint256 _fromAmount,
+        uint256 _cbltAmount
+    );
 
     /**
      * pass in the oracle contract so that it can pull info from it
@@ -49,127 +55,164 @@ contract Chromium is Ownable {
      * it will call on the oracle to make the calculation. the returnAmount is going to be
      * three times larger than the actual amount (so that we can get decimals) which means the returnAmount
      * will need to be divided by 100000 by the frontend to get the correct amount that will be swapped
-    */
+     */
     function getCbltExchangeRate(
         IERC20 fromToken,
         IERC20 destToken,
         uint256 amount
-    )
-    public
-    view
-    returns (uint returnAmount)
-    {
-        (uint256 sellTokenValue, uint256 buyTokenValue) = oracle.priceOfPair(address(fromToken), address(destToken));
-        returnAmount = SafeMath.mul(amount,
+    ) public view returns (uint256 returnAmount) {
+        (uint256 sellTokenValue, uint256 buyTokenValue) =
+            oracle.priceOfPair(address(fromToken), address(destToken));
+        returnAmount = SafeMath.mul(
+            amount,
             SafeMath.findRate(sellTokenValue, buyTokenValue)
         );
     }
 
     /**
      * @dev this function will swap cblt tokens for tokens that are allowed
-    */
-    function swapForCblt(
-        IERC20 fromToken,
-        uint256 amount
-    )
-    external
-    payable
-    {
+     */
+    function swapForCblt(IERC20 fromToken, uint256 amount) external payable {
         fromToken.universalTransferFrom(msg.sender, address(this), amount);
 
         if (fromToken == ETH_ADDRESS) {
-            uint256 usdFee =
-            SafeMath.div(
-                1000000000000000000,
-                2843
-            );
+            uint256 usdFee = SafeMath.div(1000000000000000000, 2843);
+
             amount = SafeMath.sub(amount, usdFee);
 
-            (uint256 sellTokenValue, uint256 buyTokenValue) = oracle.priceOfPair(address(fromToken), address(cbltToken));
-            uint returnAmount = SafeMath.mul(amount,
-                SafeMath.findRate(sellTokenValue, buyTokenValue)
-            );
+            (uint256 sellTokenValue, uint256 buyTokenValue) =
+                oracle.priceOfPair(address(fromToken), address(cbltToken));
+            uint256 returnAmount =
+                SafeMath.mul(
+                    amount,
+                    SafeMath.findRate(sellTokenValue, buyTokenValue)
+                );
 
             returnAmount = SafeMath.div(returnAmount, 1000000);
-            require(cbltToken.universalBalanceOf(address(this)) >= returnAmount, "Not enough tokens in Treasury.");
+            require(
+                cbltToken.universalBalanceOf(address(this)) >= returnAmount,
+                "Not enough tokens in Treasury."
+            );
 
             cbltToken.universalTransfer(msg.sender, returnAmount);
             feeTotal[fromToken] = SafeMath.add(feeTotal[fromToken], usdFee);
-            liquidityAmount[fromToken] = SafeMath.add(liquidityAmount[fromToken], SafeMath.add(amount, usdFee));
-            emit ChromiumTrade(msg.sender, address(fromToken), amount, returnAmount);
+            liquidityAmount[fromToken] = SafeMath.add(
+                liquidityAmount[fromToken],
+                SafeMath.add(amount, usdFee)
+            );
+            emit ChromiumTrade(
+                msg.sender,
+                address(fromToken),
+                amount,
+                returnAmount
+            );
         } else {
-
-            (uint256 sellTokenValue, uint256 buyTokenValue) = oracle.priceOfPair(address(fromToken), address(cbltToken));
+            (uint256 sellTokenValue, uint256 buyTokenValue) =
+                oracle.priceOfPair(address(fromToken), address(cbltToken));
 
             uint256 usdFee =
-            SafeMath.div(
-                1000000000000000000,
-                SafeMath.mul(2843, sellTokenValue)
-            );
+                SafeMath.div(
+                    1000000000000000000,
+                    SafeMath.mul(2843, sellTokenValue)
+                );
             amount = SafeMath.sub(amount, usdFee);
 
-            uint returnAmount = SafeMath.mul(amount,
-                SafeMath.findRate(sellTokenValue, buyTokenValue)
-            );
+            uint256 returnAmount =
+                SafeMath.mul(
+                    amount,
+                    SafeMath.findRate(sellTokenValue, buyTokenValue)
+                );
 
             returnAmount = SafeMath.div(returnAmount, 1000000);
-            require(cbltToken.universalBalanceOf(address(this)) >= returnAmount, "Not enough tokens in Treasury.");
+            require(
+                cbltToken.universalBalanceOf(address(this)) >= returnAmount,
+                "Not enough tokens in Treasury."
+            );
 
             cbltToken.universalTransfer(msg.sender, returnAmount);
             feeTotal[fromToken] = SafeMath.add(feeTotal[fromToken], usdFee);
-            liquidityAmount[fromToken] = SafeMath.add(liquidityAmount[fromToken], SafeMath.add(amount, usdFee));
-            emit ChromiumTrade(msg.sender, address(fromToken), amount, returnAmount);
+            liquidityAmount[fromToken] = SafeMath.add(
+                liquidityAmount[fromToken],
+                SafeMath.add(amount, usdFee)
+            );
+            emit ChromiumTrade(
+                msg.sender,
+                address(fromToken),
+                amount,
+                returnAmount
+            );
         }
-
     }
 
-    function swapCbltForToken(
-        IERC20 destToken,
-        uint amountOfCblt
-    )
-    external
-    payable
+    function swapCbltForToken(IERC20 destToken, uint256 amountOfCblt)
+        external
+        payable
     {
-        cbltToken.universalTransferFrom(msg.sender, address(this), amountOfCblt);
+        cbltToken.universalTransferFrom(
+            msg.sender,
+            address(this),
+            amountOfCblt
+        );
 
         if (destToken == ETH_ADDRESS) {
-            (uint256 sellTokenValue, uint256 buyTokenValue) = oracle.priceOfPair(address(cbltToken), address(destToken));
-            uint returnAmount = SafeMath.mul(amountOfCblt,
-                SafeMath.findRate(sellTokenValue, buyTokenValue)
-            );
+            (uint256 sellTokenValue, uint256 buyTokenValue) =
+                oracle.priceOfPair(address(cbltToken), address(destToken));
+            uint256 returnAmount =
+                SafeMath.mul(
+                    amountOfCblt,
+                    SafeMath.findRate(sellTokenValue, buyTokenValue)
+                );
 
             uint256 usdFee =
-            SafeMath.div(
-                1000000000000000000,
-                SafeMath.mul(2843, buyTokenValue)
-            );
+                SafeMath.div(
+                    1000000000000000000,
+                    SafeMath.mul(2843, buyTokenValue)
+                );
 
             returnAmount = SafeMath.sub(returnAmount, usdFee);
             returnAmount = SafeMath.div(returnAmount, 1000000);
-            require(destToken.universalBalanceOf(address(this)) >= returnAmount, "Not enough tokens in Treasury.");
+            require(
+                destToken.universalBalanceOf(address(this)) >= returnAmount,
+                "Not enough tokens in Treasury."
+            );
 
             destToken.universalTransfer(msg.sender, returnAmount);
-            emit ChromiumTrade(msg.sender, address(cbltToken), amountOfCblt, returnAmount);
+            emit ChromiumTrade(
+                msg.sender,
+                address(cbltToken),
+                amountOfCblt,
+                returnAmount
+            );
         } else {
-
-            (uint256 sellTokenValue, uint256 buyTokenValue) = oracle.priceOfPair(address(cbltToken), address(destToken));
-            uint returnAmount = SafeMath.mul(amountOfCblt,
-                SafeMath.findRate(sellTokenValue, buyTokenValue)
-            );
+            (uint256 sellTokenValue, uint256 buyTokenValue) =
+                oracle.priceOfPair(address(cbltToken), address(destToken));
+            uint256 returnAmount =
+                SafeMath.mul(
+                    amountOfCblt,
+                    SafeMath.findRate(sellTokenValue, buyTokenValue)
+                );
 
             uint256 usdFee =
-            SafeMath.div(
-                1000000000000000000,
-                SafeMath.mul(1000000000000000000, 2843)
-            );
+                SafeMath.div(
+                    1000000000000000000,
+                    SafeMath.mul(1000000000000000000, 2843)
+                );
 
             returnAmount = SafeMath.sub(returnAmount, usdFee);
             returnAmount = SafeMath.div(returnAmount, 1000000);
 
-            require(destToken.universalBalanceOf(address(this)) >= returnAmount, "Not enough tokens in Treasury.");
+            require(
+                destToken.universalBalanceOf(address(this)) >= returnAmount,
+                "Not enough tokens in Treasury."
+            );
             destToken.universalTransfer(msg.sender, returnAmount);
 
-            emit ChromiumTrade(msg.sender, address(cbltToken), amountOfCblt, returnAmount);
+            emit ChromiumTrade(
+                msg.sender,
+                address(cbltToken),
+                amountOfCblt,
+                returnAmount
+            );
         }
     }
